% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firthGP.R
\name{firthGP}
\alias{firthGP}
\title{Estimate generalized Poisson regression model for count data where the data has separation, quasi-separation or near-to-quasi separation problem}
\usage{
firthGP(formula, data = parent.frame(), alpha = 0.05)
}
\arguments{
\item{formula}{similar as glm}

\item{data}{a data frame or tibble}

\item{alpha}{level of significance for confidence interval}
}
\value{
a table of coefficients and ci
}
\description{
Estimate generalized Poisson regression model for count data where the data has separation, quasi-separation or near-to-quasi separation problem
}
\examples{

quasifit = firthGP(y ~ x1 + x2, data = quasiSep)
quasifit





}
