% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearSep.R
\docType{data}
\name{nearSep}
\alias{nearSep}
\title{A simulated count data that has near-to-quasi-separation problem}
\format{
\subsection{\code{nearSep}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{y}{the response variable, count data}
\item{x1}{a continuous variable}
\item{x2}{a binary variable that causes near-to-quasi-separation problem}
}
}
}
\usage{
nearSep
}
\description{
A simulated count data that has near-to-quasi-separation problem
}
\keyword{datasets}
