% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiDat.R
\docType{data}
\name{quasiDat}
\alias{quasiDat}
\title{Antenatal care (ANC) data from BDHS 2017-2018
A total 100 mothers randomly (seed 323) selected from total adolescent mothers
(age < 20)}
\format{
\subsection{\code{quasiDat}}{

A data frame with 100 rows and 8 columns:
\describe{
\item{anc_visit}{number of ANC visits were made during her pregnancy}
\item{educ_level}{mother's educational status, educated or uneducated}
\item{media}{exposed to media, 0 = non-exposed, 1 = exposed}
\item{area}{geographical area she lives in}
\item{division}{geographical division she lives in}
\item{wi}{wealth index of five groups: poorest, poorer, middle, richer, richest}
\item{husb_edu}{husband's education levels}
\item{age}{mother's age}

}
}
}
\usage{
quasiDat
}
\description{
Antenatal care (ANC) data from BDHS 2017-2018
A total 100 mothers randomly (seed 323) selected from total adolescent mothers
(age < 20)
}
\keyword{datasets}
