% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiSep.R
\docType{data}
\name{quasiSep}
\alias{quasiSep}
\title{A simulated count data that has quasi-separation problem}
\format{
\subsection{\code{quasiSep}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{y}{the response variable, count data}
\item{x1}{a continuous variable}
\item{x2}{a binary variable that causes quasi-separation problem}
}
}
}
\usage{
quasiSep
}
\description{
A simulated count data that has quasi-separation problem
}
\keyword{datasets}
